feat: extracted expiry logic from the GenericReadableMessage to a new component
This allows use to use the expiry logic for different visible messages not just generic ones
